requirements.txt
fastapi==0.75.1
uvicorn==0.17.6
sqlalchemy==1.4.36
psycopg2-binary==2.9.3
fastapi-pagination==0.9.3
pydantic==1.9.1
python-decouple==3.6
pre-commit==2.20.0
black==22.6.0
flake8==4.0.1